requirements.txt
Adafruit-Blinka==6.15.0
adafruit-circuitpython-neopixel==6.2.3
adafruit-circuitpython-pixelbuf==1.1.1
Adafruit-PlatformDetect==3.17.2
Adafruit-PureIO==1.1.9
anyio==3.4.0
certifi==2021.10.8
charset-normalizer==2.0.7
h11==0.12.0
httpcore==0.14.3
httpx==0.21.1
idna==3.3
pyftdi==0.53.3
pyserial==3.5
pyusb==1.2.1
rfc3986==1.5.0
sniffio==1.2.0
astral==2.2
